Se finalizan las paginas usuarios e ingredientes, pedidos a corregir el script de la tabla, falta desarrollo en la pagina de mesas y comidas. Se finaliza el desarrollo de la aplicacion de escritorio

+++ TPC_Muñiz/TPC_Muñiz/Managers/StockManager.cs
@@ -76,7 +76,7 @@ namespace TPC_Muñiz.Managers
         }
 
 
-        public BindingList<Ingrediente> Cargarstocks(int id, int cant)
+        public BindingList<Ingrediente> Cargarstocks(int id, float cant)
         {
             using (var cn = new SqlConnection(getConnectionString()))
             {
